Revert "Fix pending notifications not being received after going online"
